Tidy up class structure and docs
