+++ insets/src/main/java/com/google/accompanist/insets/Insets.kt
@@ -14,29 +14,20 @@
  * limitations under the License.
  */
 
-@file:Suppress("NOTHING_TO_INLINE", "unused", "PropertyName")
-
-@file:JvmName("ComposeInsets")
-@file:JvmMultifileClass
-
 package com.google.accompanist.insets
 
-import android.view.View
-import android.view.WindowInsetsAnimation
 import androidx.annotation.IntRange
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.Stable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.platform.LocalView
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsAnimationCompat
-import androidx.core.view.WindowInsetsCompat
+import com.google.accompanist.insets.Insets.Companion.Insets
 
+/**
+ * Interface which represents a single set of insets. Each instance holds four integer offsets
+ * which describe changes to the four edges of a rectangle.
+ */
 @Stable
 interface Insets {
     /**
@@ -63,6 +54,9 @@ interface Insets {
     @get:IntRange(from = 0)
     val bottom: Int
 
+    /**
+     * Returns a copy of this instance with the given values.
+     */
     fun copy(
         left: Int = this.left,
         top: Int = this.top,
@@ -85,6 +79,9 @@ interface Insets {
     )
 
     companion object {
+        /**
+         * Creates an [Insets] instance with the given values.
+         */
         fun Insets(
             left: Int = 0,
             top: Int = 0,
@@ -92,17 +89,27 @@ interface Insets {
             bottom: Int = 0,
         ): Insets = ImmutableInsets(left, top, right, bottom)
 
+        /**
+         * An empty [Insets] instance, with each dimension set to a value of 0.
+         */
         val Empty: Insets = ImmutableInsets()
     }
 }
 
-private class ImmutableInsets(
+/**
+ * Immutable implementation of [Insets].
+ */
+@Immutable
+internal class ImmutableInsets(
     override val left: Int = 0,
     override val top: Int = 0,
     override val right: Int = 0,
     override val bottom: Int = 0,
 ) : Insets
 
+/**
+ * Mutable [androidx.compose.runtime.State] backed implementation of [Insets].
+ */
 internal class MutableInsets(
     left: Int = 0,
     top: Int = 0,
@@ -130,369 +137,11 @@ internal class MutableInsets(
 }
 
 /**
- * This class sets up the necessary listeners on the given [view] to be able to observe
- * [WindowInsetsCompat] instances dispatched by the system.
- *
- * This class is useful for when you prefer to handle the ownership of the [WindowInsets]
- * yourself. One example of this is if you find yourself using [ProvideWindowInsets] in fragments.
- *
- * It is convenient to use [ProvideWindowInsets] in fragments, but that can result in a
- * delay in the initial inset update, which results in a visual flicker.
- * See [this issue](https://github.com/google/accompanist/issues/155) for more information.
- *
- * The alternative is for fragments to manage the [WindowInsets] themselves, like so:
- *
- * ```
- * override fun onCreateView(
- *     inflater: LayoutInflater,
- *     container: ViewGroup?,
- *     savedInstanceState: Bundle?
- * ): View = ComposeView(requireContext()).apply {
- *     layoutParams = LayoutParams(MATCH_PARENT, MATCH_PARENT)
- *
- *     // Create an ViewWindowInsetObserver using this view
- *     val observer = ViewWindowInsetObserver(this)
- *
- *     // Call start() to start listening now.
- *     // The WindowInsets instance is returned to us.
- *     val windowInsets = observer.start()
- *
- *     setContent {
- *         // Instead of calling ProvideWindowInsets, we use CompositionLocalProvider to provide
- *         // the WindowInsets instance from above to LocalWindowInsets
- *         CompositionLocalProvider(LocalWindowInsets provides windowInsets) {
- *             /* Content */
- *         }
- *     }
- * }
- * ```
- *
- * @param view The view to observe [WindowInsetsCompat]s from.
- */
-class ViewWindowInsetObserver(private val view: View) {
-    private val attachListener = object : View.OnAttachStateChangeListener {
-        override fun onViewAttachedToWindow(v: View) = v.requestApplyInsets()
-        override fun onViewDetachedFromWindow(v: View) = Unit
-    }
-
-    /**
-     * Whether this [ViewWindowInsetObserver] is currently observing.
-     */
-    @Suppress("MemberVisibilityCanBePrivate")
-    var isObserving: Boolean = false
-        private set
-
-    /**
-     * Start observing window insets from [view]. Make sure to call [stop] if required.
-     *
-     * @param consumeWindowInsets Whether to consume any [WindowInsetsCompat]s which are
-     * dispatched to the host view. Defaults to `true`.
-     */
-    fun start(
-        consumeWindowInsets: Boolean = true
-    ): WindowInsets = RootWindowInsets().also {
-        observeInto(
-            windowInsets = it,
-            consumeWindowInsets = consumeWindowInsets,
-            windowInsetsAnimationsEnabled = false
-        )
-    }
-
-    /**
-     * Start observing window insets from [view]. Make sure to call [stop] if required.
-     *
-     * @param windowInsetsAnimationsEnabled Whether to listen for [WindowInsetsAnimation]s, such as
-     * IME animations.
-     * @param consumeWindowInsets Whether to consume any [WindowInsetsCompat]s which are
-     * dispatched to the host view. Defaults to `true`.
-     */
-    @ExperimentalAnimatedInsets
-    fun start(
-        windowInsetsAnimationsEnabled: Boolean,
-        consumeWindowInsets: Boolean = true,
-    ): WindowInsets = RootWindowInsets().also {
-        observeInto(
-            windowInsets = it,
-            consumeWindowInsets = consumeWindowInsets,
-            windowInsetsAnimationsEnabled = windowInsetsAnimationsEnabled
-        )
-    }
-
-    internal fun observeInto(
-        windowInsets: RootWindowInsets,
-        consumeWindowInsets: Boolean,
-        windowInsetsAnimationsEnabled: Boolean,
-    ) {
-        require(!isObserving) {
-            "start() called, but this ViewWindowInsetObserver is already observing"
-        }
-
-        ViewCompat.setOnApplyWindowInsetsListener(view) { _, wic ->
-            // Go through each inset type and update its layoutInsets from the
-            // WindowInsetsCompat values
-            windowInsets.statusBars.run {
-                _layoutInsets.updateFrom(wic.getInsets(WindowInsetsCompat.Type.statusBars()))
-                isVisible = wic.isVisible(WindowInsetsCompat.Type.statusBars())
-            }
-            windowInsets.navigationBars.run {
-                _layoutInsets.updateFrom(wic.getInsets(WindowInsetsCompat.Type.navigationBars()))
-                isVisible = wic.isVisible(WindowInsetsCompat.Type.navigationBars())
-            }
-            windowInsets.systemGestures.run {
-                _layoutInsets.updateFrom(wic.getInsets(WindowInsetsCompat.Type.systemGestures()))
-                isVisible = wic.isVisible(WindowInsetsCompat.Type.systemGestures())
-            }
-            windowInsets.ime.run {
-                _layoutInsets.updateFrom(wic.getInsets(WindowInsetsCompat.Type.ime()))
-                isVisible = wic.isVisible(WindowInsetsCompat.Type.ime())
-            }
-
-            if (consumeWindowInsets) WindowInsetsCompat.CONSUMED else wic
-        }
-
-        // Add an OnAttachStateChangeListener to request an inset pass each time we're attached
-        // to the window
-        val attachListener = object : View.OnAttachStateChangeListener {
-            override fun onViewAttachedToWindow(v: View) = v.requestApplyInsets()
-            override fun onViewDetachedFromWindow(v: View) = Unit
-        }
-        view.addOnAttachStateChangeListener(attachListener)
-
-        if (windowInsetsAnimationsEnabled) {
-            ViewCompat.setWindowInsetsAnimationCallback(
-                view,
-                InnerWindowInsetsAnimationCallback(windowInsets)
-            )
-        } else {
-            ViewCompat.setWindowInsetsAnimationCallback(view, null)
-        }
-
-        if (view.isAttachedToWindow) {
-            // If the view is already attached, we can request an inset pass now
-            view.requestApplyInsets()
-        }
-
-        isObserving = true
-    }
-
-    /**
-     * Removes any listeners from the [view] so that we no longer observe inset changes.
-     *
-     * This is only required to be called from hosts which have a shorter lifetime than the [view].
-     * For example, if you're using [ViewWindowInsetObserver] from a `@Composable` function,
-     * you should call [stop] from an `onDispose` block, like so:
-     *
-     * ```
-     * DisposableEffect(view) {
-     *     val observer = ViewWindowInsetObserver(view)
-     *     // ...
-     *     onDispose {
-     *         observer.stop()
-     *     }
-     * }
-     * ```
-     *
-     * Whereas if you're using this class from a fragment (or similar), it is not required to
-     * call this function since it will live as least as longer as the view.
-     */
-    fun stop() {
-        require(isObserving) {
-            "stop() called, but this ViewWindowInsetObserver is not currently observing"
-        }
-        view.removeOnAttachStateChangeListener(attachListener)
-        ViewCompat.setOnApplyWindowInsetsListener(view, null)
-        isObserving = false
-    }
-}
-
-/**
- * Applies any [WindowInsetsCompat] values to [LocalWindowInsets], which are then available
- * within [content].
- *
- * If you're using this in fragments, you may wish to take a look at
- * [ViewWindowInsetObserver] for a more optimal solution.
- *
- * @param consumeWindowInsets Whether to consume any [WindowInsetsCompat]s which are dispatched to
- * the host view. Defaults to `true`.
- */
-@Composable
-fun ProvideWindowInsets(
-    consumeWindowInsets: Boolean = true,
-    content: @Composable () -> Unit,
-) {
-    val view = LocalView.current
-    val windowInsets = RootWindowInsets()
-
-    DisposableEffect(view) {
-        val observer = ViewWindowInsetObserver(view)
-        observer.observeInto(
-            windowInsets = windowInsets,
-            consumeWindowInsets = consumeWindowInsets,
-            windowInsetsAnimationsEnabled = false
-        )
-        onDispose {
-            observer.stop()
-        }
-    }
-
-    CompositionLocalProvider(LocalWindowInsets provides windowInsets) {
-        content()
-    }
-}
-
-/**
- * Applies any [WindowInsetsCompat] values to [LocalWindowInsets], which are then available
- * within [content].
- *
- * If you're using this in fragments, you may wish to take a look at
- * [ViewWindowInsetObserver] for a more optimal solution.
- *
- * @param windowInsetsAnimationsEnabled Whether to listen for [WindowInsetsAnimation]s, such as
- * IME animations.
- * @param consumeWindowInsets Whether to consume any [WindowInsetsCompat]s which are dispatched to
- * the host view. Defaults to `true`.
- */
-@ExperimentalAnimatedInsets
-@Composable
-fun ProvideWindowInsets(
-    windowInsetsAnimationsEnabled: Boolean,
-    consumeWindowInsets: Boolean = true,
-    content: @Composable () -> Unit
-) {
-    val view = LocalView.current
-    val windowInsets = remember { RootWindowInsets() }
-
-    DisposableEffect(view) {
-        val observer = ViewWindowInsetObserver(view)
-        observer.observeInto(
-            windowInsets = windowInsets,
-            consumeWindowInsets = consumeWindowInsets,
-            windowInsetsAnimationsEnabled = windowInsetsAnimationsEnabled
-        )
-        onDispose { observer.stop() }
-    }
-
-    CompositionLocalProvider(LocalWindowInsets provides windowInsets) {
-        content()
-    }
-}
-
-private class InnerWindowInsetsAnimationCallback(
-    private val windowInsets: RootWindowInsets,
-) : WindowInsetsAnimationCompat.Callback(DISPATCH_MODE_STOP) {
-    override fun onPrepare(animation: WindowInsetsAnimationCompat) {
-        // Go through each type and flag that an animation has started
-        if (animation.typeMask and WindowInsetsCompat.Type.ime() != 0) {
-            windowInsets.ime.onAnimationStart()
-        }
-        if (animation.typeMask and WindowInsetsCompat.Type.statusBars() != 0) {
-            windowInsets.statusBars.onAnimationStart()
-        }
-        if (animation.typeMask and WindowInsetsCompat.Type.navigationBars() != 0) {
-            windowInsets.navigationBars.onAnimationStart()
-        }
-        if (animation.typeMask and WindowInsetsCompat.Type.systemGestures() != 0) {
-            windowInsets.systemGestures.onAnimationStart()
-        }
-    }
-
-    override fun onProgress(
-        platformInsets: WindowInsetsCompat,
-        runningAnimations: List<WindowInsetsAnimationCompat>
-    ): WindowInsetsCompat {
-        // Update each inset type with the given parameters
-        windowInsets.ime.updateAnimation(
-            platformInsets = platformInsets,
-            runningAnimations = runningAnimations,
-            type = WindowInsetsCompat.Type.ime()
-        )
-        windowInsets.statusBars.updateAnimation(
-            platformInsets = platformInsets,
-            runningAnimations = runningAnimations,
-            type = WindowInsetsCompat.Type.statusBars()
-        )
-        windowInsets.navigationBars.updateAnimation(
-            platformInsets = platformInsets,
-            runningAnimations = runningAnimations,
-            type = WindowInsetsCompat.Type.navigationBars()
-        )
-        windowInsets.systemGestures.updateAnimation(
-            platformInsets = platformInsets,
-            runningAnimations = runningAnimations,
-            type = WindowInsetsCompat.Type.systemGestures()
-        )
-        return platformInsets
-    }
-
-    private inline fun MutableInsetsType.updateAnimation(
-        platformInsets: WindowInsetsCompat,
-        runningAnimations: List<WindowInsetsAnimationCompat>,
-        type: Int,
-    ) {
-        // If there are animations of the given type...
-        if (runningAnimations.any { it.typeMask or type != 0 }) {
-            // Update our animated inset values
-            _animatedInsets.updateFrom(platformInsets.getInsets(type))
-            // And update the animation fraction. We use the maximum animation progress of any
-            // ongoing animations for this type.
-            animationFraction = runningAnimations.maxOf { it.fraction }
-        }
-    }
-
-    override fun onEnd(animation: WindowInsetsAnimationCompat) {
-        // Go through each type and flag that an animation has ended
-        if (animation.typeMask and WindowInsetsCompat.Type.ime() != 0) {
-            windowInsets.ime.onAnimationEnd()
-        }
-        if (animation.typeMask and WindowInsetsCompat.Type.statusBars() != 0) {
-            windowInsets.statusBars.onAnimationEnd()
-        }
-        if (animation.typeMask and WindowInsetsCompat.Type.navigationBars() != 0) {
-            windowInsets.navigationBars.onAnimationEnd()
-        }
-        if (animation.typeMask and WindowInsetsCompat.Type.systemGestures() != 0) {
-            windowInsets.systemGestures.onAnimationEnd()
-        }
-    }
-}
-
-/**
  * Updates our mutable state backed [InsetsType] from an Android system insets.
  */
-private fun MutableInsets.updateFrom(insets: androidx.core.graphics.Insets) {
+internal fun MutableInsets.updateFrom(insets: androidx.core.graphics.Insets) {
     left = insets.left
     top = insets.top
     right = insets.right
     bottom = insets.bottom
 }
-
-/**
- * Ensures that each dimension is not less than corresponding dimension in the
- * specified [minimumValue].
- *
- * @return this if every dimension is greater than or equal to the corresponding
- * dimension value in [minimumValue], otherwise a copy of this with each dimension coerced with the
- * corresponding dimension value in [minimumValue].
- */
-fun InsetsType.coerceEachDimensionAtLeast(minimumValue: InsetsType): Insets {
-    // Fast path, no need to copy if: this >= minimumValue
-    if (left >= minimumValue.left && top >= minimumValue.top &&
-        right >= minimumValue.right && bottom >= minimumValue.bottom
-    ) {
-        return this
-    }
-    return MutableInsets(
-        left = left.coerceAtLeast(minimumValue.left),
-        top = top.coerceAtLeast(minimumValue.top),
-        right = right.coerceAtLeast(minimumValue.right),
-        bottom = bottom.coerceAtLeast(minimumValue.bottom),
-    )
-}
-
-enum class HorizontalSide { Left, Right }
-enum class VerticalSide { Top, Bottom }
-
-@RequiresOptIn(message = "Animated Insets support is experimental. The API may be changed in the future.")
-@Retention(AnnotationRetention.BINARY)
-@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION)
-annotation class ExperimentalAnimatedInsets
